Conflicts:
	src/CMake/FindOptiX.cmake
	src/CMakeLists.txt
	src/ElVis/Extensions/NektarPlusPlusExtension/CMakeLists.txt

+++ src/ElVis/Core/Color.cpp
@@ -66,20 +66,20 @@ namespace ElVis
     float Color::Blue() const { return m_blue; }
     float Color::Alpha() const { return m_alpha; }
 
-    unsigned int Color::RedAsInt() const { return static_cast<unsigned int>(m_red*255); }
-    unsigned int Color::GreenAsInt() const { return static_cast<unsigned int>(m_green*255); }
-    unsigned int Color::BlueAsInt() const { return static_cast<unsigned int>(m_blue*255); }
-    unsigned int Color::AlphaAsInt() const { return static_cast<unsigned int>(m_alpha*255); }
+    int Color::RedAsInt() const { return static_cast<int>(m_red*255); }
+    int Color::GreenAsInt() const { return static_cast<int>(m_green*255); }
+    int Color::BlueAsInt() const { return static_cast<int>(m_blue*255); }
+    int Color::AlphaAsInt() const { return static_cast<int>(m_alpha*255); }
 
     void Color::SetRed(double value) { m_red = static_cast<float>(value); OnColorChanged(*this); }
     void Color::SetGreen(double value) { m_green = static_cast<float>(value); OnColorChanged(*this); }
     void Color::SetBlue(double value) { m_blue = static_cast<float>(value); OnColorChanged(*this); }
     void Color::SetAlpha(double value) { m_alpha = static_cast<float>(value); OnColorChanged(*this);}
 
-    void Color::SetRed(int value) { m_red = (float)value/255.0; OnColorChanged(*this);}
-    void Color::SetGreen(int value) { m_green = (float)value/255.0; OnColorChanged(*this);}
-    void Color::SetBlue(int value) { m_blue = (float)value/255.0; OnColorChanged(*this);}
-    void Color::SetAlpha(int value) { m_alpha = (float)value/255.0; OnColorChanged(*this);}
+    void Color::SetRed(int value) { m_red = value/255.0f; OnColorChanged(*this);}
+    void Color::SetGreen(int value) { m_green = value/255.0f; OnColorChanged(*this);}
+    void Color::SetBlue(int value) { m_blue = value/255.0f; OnColorChanged(*this);}
+    void Color::SetAlpha(int value) { m_alpha = value/255.0f; OnColorChanged(*this);}
 
     Color operator+(const Color& lhs, const Color& rhs)
     {
